requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6